Refactoring, moving some classes to different packages
