+++ src/main/java/com/safetynet/alerts/logic/GetService.java
@@ -3,6 +3,9 @@ package com.safetynet.alerts.logic;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import com.safetynet.alerts.logging.LogHandlerTiny;
+import com.safetynet.alerts.logic.parsers.CollectionParser;
+import com.safetynet.alerts.logic.parsers.ModelObjectFinder;
+import com.safetynet.alerts.logic.parsers.PersonAndRecordParser;
 import com.safetynet.alerts.presentation.model.entity.*;
 import com.safetynet.alerts.presentation.model.*;
 import org.springframework.http.HttpHeaders;
