+++ src/main/java/com/safetynet/alerts/logic/updaters/ResultModel.java
@@ -1,4 +1,4 @@
-package com.safetynet.alerts.logic;
+package com.safetynet.alerts.logic.updaters;
 
 import com.safetynet.alerts.presentation.model.SafetyAlertsModel;
