+++ src/main/java/com/safetynet/alerts/presentation/controller/GetMappingController.java
@@ -4,6 +4,9 @@ import com.safetynet.alerts.configuration.DataConfig;
 import com.safetynet.alerts.data.io.JsonDAO;
 import com.safetynet.alerts.logging.LogHandlerTiny;
 import com.safetynet.alerts.logic.*;
+import com.safetynet.alerts.logic.parsers.CollectionParser;
+import com.safetynet.alerts.logic.parsers.ModelObjectFinder;
+import com.safetynet.alerts.logic.parsers.PersonAndRecordParser;
 import com.safetynet.alerts.presentation.model.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
