+++ src/main/java/com/safetynet/alerts/presentation/controller/MedicalRecordController.java
@@ -5,9 +5,9 @@ import com.google.gson.GsonBuilder;
 import com.safetynet.alerts.configuration.DataConfig;
 import com.safetynet.alerts.data.io.JsonDAO;
 import com.safetynet.alerts.logging.LogHandlerTiny;
-import com.safetynet.alerts.logic.ModelObjectFinder;
-import com.safetynet.alerts.logic.ResultModel;
-import com.safetynet.alerts.logic.UpdateMedicalRecord;
+import com.safetynet.alerts.logic.parsers.ModelObjectFinder;
+import com.safetynet.alerts.logic.updaters.ResultModel;
+import com.safetynet.alerts.logic.updaters.UpdateMedicalRecord;
 import com.safetynet.alerts.logic.JsonHandler;
 import com.safetynet.alerts.presentation.model.MedicalRecord;
 import com.safetynet.alerts.presentation.model.SafetyAlertsModel;
