+++ src/test/java/com/safetynet/alerts/logic/UpdateFirestationTest.java
@@ -1,5 +1,8 @@
 package com.safetynet.alerts.logic;
 
+import com.safetynet.alerts.logic.parsers.ModelObjectFinder;
+import com.safetynet.alerts.logic.updaters.ResultModel;
+import com.safetynet.alerts.logic.updaters.UpdateFirestation;
 import com.safetynet.alerts.presentation.model.Firestation;
 import com.safetynet.alerts.presentation.model.MedicalRecord;
 import com.safetynet.alerts.presentation.model.Person;
