+++ src/test/java/com/safetynet/alerts/presentation/controller/DataFileErrorTest.java
@@ -5,6 +5,12 @@ import com.safetynet.alerts.data.io.JsonDAO;
 import com.safetynet.alerts.logging.LogHandlerTiny;
 import com.safetynet.alerts.logic.*;
 import com.safetynet.alerts.logic.JsonHandler;
+import com.safetynet.alerts.logic.parsers.CollectionParser;
+import com.safetynet.alerts.logic.parsers.ModelObjectFinder;
+import com.safetynet.alerts.logic.parsers.PersonAndRecordParser;
+import com.safetynet.alerts.logic.updaters.UpdateFirestation;
+import com.safetynet.alerts.logic.updaters.UpdateMedicalRecord;
+import com.safetynet.alerts.logic.updaters.UpdatePerson;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
