+++ src/test/java/com/safetynet/alerts/presentation/controller/GetControllerBeanSetup.java
@@ -3,10 +3,10 @@ package com.safetynet.alerts.presentation.controller;
 import com.safetynet.alerts.configuration.DataConfig;
 import com.safetynet.alerts.data.io.JsonDAO;
 import com.safetynet.alerts.logging.LogHandlerTiny;
-import com.safetynet.alerts.logic.CollectionParser;
+import com.safetynet.alerts.logic.parsers.CollectionParser;
 import com.safetynet.alerts.logic.GetService;
-import com.safetynet.alerts.logic.ModelObjectFinder;
-import com.safetynet.alerts.logic.PersonAndRecordParser;
+import com.safetynet.alerts.logic.parsers.ModelObjectFinder;
+import com.safetynet.alerts.logic.parsers.PersonAndRecordParser;
 import com.safetynet.alerts.logic.JsonHandler;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.context.annotation.Bean;
