+++ src/test/java/com/safetynet/alerts/presentation/controller/ObjectControllerBeanSetup.java
@@ -3,8 +3,11 @@ package com.safetynet.alerts.presentation.controller;
 import com.safetynet.alerts.configuration.DataConfig;
 import com.safetynet.alerts.data.io.JsonDAO;
 import com.safetynet.alerts.logging.LogHandlerTiny;
-import com.safetynet.alerts.logic.*;
 import com.safetynet.alerts.logic.JsonHandler;
+import com.safetynet.alerts.logic.parsers.ModelObjectFinder;
+import com.safetynet.alerts.logic.updaters.UpdateFirestation;
+import com.safetynet.alerts.logic.updaters.UpdateMedicalRecord;
+import com.safetynet.alerts.logic.updaters.UpdatePerson;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.context.annotation.Bean;
